MessagingGroupsConsumer base structure.
