+++ messaging-groups-consumer/bin/install.sh
@@ -4,13 +4,13 @@
 ##
 
 # Assume messagebroker-config repo is one directory up
-cd ../messagebroker-config
+cd ../../messagebroker-config
 
 # Gather path from root
 MBCONFIG=`pwd`
 
 # Back to mbp-user-import
-cd ../mbc-a1-startHere
+cd ../MessageBroker-PHP/messaging-groups-consumer
 
 # Create SymLink for mbp-user-import application to make reference to for all Message Broker configuration settings
 ln -s $MBCONFIG .
